chore: 프로젝트 그룹 정리
