+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/TaskManager.swift
@@ -62,10 +62,3 @@ extension TaskManager: ClientEnqueuable {
         self.clientQueue.enqueue(client)
     }
 }
-
-extension TaskManager: StatGeneratable {
-    func generateStat() -> TaskStat {
-        // TODO: 결과 집계 구현 (리턴..)
-        return TaskStat(clientCount: 0)
-    }
-}
