+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/TicketDispenser.swift
