+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/Value/Client.swift
