--- BankManagerConsoleApp/BankManagerConsoleApp/Protocol/StatGeneratable.swift
@@ -1,10 +1,0 @@
-//
-//  StatGeneratable.swift
-//  BankManagerConsoleApp
-//
-//  Created by Effie on 1/31/24.
-//
-
-protocol StatGeneratable {
-    func generateStat() -> TaskStat
-}
