--- BankManagerConsoleApp/BankManagerConsoleApp/TaskStat.swift
@@ -1,10 +1,0 @@
-//
-//  TaskStat.swift
-//  BankManagerConsoleApp
-//
-//  Created by Effie on 1/31/24.
-//
-
-struct TaskStat {
-    let clientCount: Int
-}
