+++ BankManagerConsoleApp/BankManagerConsoleApp/Utils/LinkedList.swift
@@ -1,19 +1,9 @@
 //
-//  Queue.swift
+//  LinkedList.swift
 //  BankManagerConsoleApp
 //
-//  Created by Effie on 1/31/24.
+//  Created by 강창현 on 2/1/24.
 //
-
-final class Node<Value> {
-    let value: Value
-    var next: Node?
-    
-    init(value: Value, next: Node? = nil) {
-        self.value = value
-        self.next = next
-    }
-}
 
 final class LinkedList<Value> {
     private(set) var head: Node<Value>?
@@ -72,46 +62,5 @@ final class LinkedList<Value> {
         self.head = nil
         self.tail = nil
         self.count = 0
-    }
-}
-
-final class Queue<Value> {
-    private var linkedList: LinkedList<Value>
-    
-    var front: Node<Value>? {
-        return self.linkedList.head
-    }
-    
-    var rear: Node<Value>? {
-        return self.linkedList.tail
-    }
-    
-    var isEmpty: Bool {
-        return self.linkedList.isEmpty
-    }
-    
-    var count: Int {
-        return self.linkedList.count
-    }
-    
-    init(linkedList: LinkedList<Value> = LinkedList()) {
-        self.linkedList = linkedList
-    }
-    
-    func enqueue(_ value: Value) {
-        self.linkedList.add(value: value)
-    }
-    
-    @discardableResult
-    func dequeue() -> Value? {
-        return self.linkedList.removeFirst()
-    }
-    
-    func clear() {
-        self.linkedList.clear()
-    }
-    
-    func peek() -> Value? {
-        return self.linkedList.first
     }
 }
