+++ BankManagerConsoleApp/BankManagerConsoleApp/View/ConsoleManager.swift
