+++ BankManagerConsoleApp/BankManagerConsoleApp/main.swift
@@ -3,5 +3,6 @@
 //  Created by yagom.
 //  Copyright © yagom academy. All rights reserved.
 //
+
 let console = ConsoleManager()
 BankManagerApp(inputHandler: console, outputHandler: console).start()
